pyproject.toml
[build-system]
requires = ["flit_core >=2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "logDet"
version = "0.0.1"
description = "Anomaly detection for log data"
authors = [
    { name = "Amine MEKKI", email = "amine.mekki.contact@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "numpy",
    "torch",
    "scikit-learn",
    "matplotlib",
    "python-box",
    "pyYaml",
    "types-pyYaml",
    "ensure"
]
[project.optional-dependencies]
dev = [
    "ruff",
    "black",
]
